+++ h6x_serial_interface/include/h6x_serial_interface/h6x_serial_interface.hpp
@@ -16,3 +16,4 @@
 
 #include "h6x_serial_interface/port_handler.hpp"
 #include "h6x_serial_interface/port_handler_base.hpp"
+#include "h6x_serial_interface/serial_socket.hpp"
