--- h6x_serial_interface/include/h6x_serial_interface/socket_handler_impl.hpp
@@ -1,19 +1,0 @@
-/*
- * Copyright (c) 2024 HarvestX Inc.
- *
- * SPDX-License-Identifier: Apache-2.0
- */
-
-#include "socket_handler.hpp"
-
-namespace h6x_serial_interface
-{
-
-bool SerialSocket::configure(const int baud = 115200)
-{
-  using boost::asio::serial_port_base;
-  try {
-    socket_->
-  }
-}
-};  // namespace h6x_serial_interface
